requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.0.0
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
awscli==2.9.19
awscrt==1.0.0.dev0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==6.0.0
blinker==1.5
Brotli==1.0.9
certifi==2023.7.22
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dm-tree==0.1.8
docutils==0.19
eagerpy==0.30.0
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2023.12.2
gast==0.5.4
google-pasta==0.2.0
grpcio==1.60.0
gudhi==3.7.1
h5py==3.10.0
html5lib==1.1
httplib2==0.20.4
idna==3.4
importlib-metadata==7.0.1
iniconfig==1.1.1
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras-nightly==3.0.3.dev2024011803
keyring==24.3.0
kiwisolver==0.0.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libclang==16.0.6
lxml==4.9.2
lz4==4.0.2+dfsg
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.7
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nh3==0.2.15
nose==1.3.7
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.0.0+repack
ply==3.11
POT==0.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pythran==0.11.0
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.1.1
readme-renderer==42.0
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
roman==3.3
rpds-py==0.10.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.4.2
scipy==1.13.0
SecretStorage==3.3.3
Send2Trash==1.8.2
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tb-nightly==2.16.0a20240118
tensorboard-data-server==0.7.2
tensorflow==2.16.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.17.1
tf-estimator-nightly==2.14.0.dev2023080308
tf_keras-nightly==2.16.0.dev2024011811
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
torch==2.1.2
torch_geometric==2.5.3
tornado==6.3.3
tqdm==4.66.2
traitlets==5.11.1
triton==2.1.0
twine==4.0.2
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2025.1
ufoLib2==0.14.0
uri-template==1.3.0
urllib3==1.26.12
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.1
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.9.4
yq==3.1.0
zipp==3.17.0
